ai-agent/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
anthropic>=0.40.0
httpx>=0.28.0
pydantic>=2.10.3
pydantic-settings>=2.7.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
rich>=13.9.4
typer>=0.15.1
psycopg2-binary>=2.9.10
# Jarvis multi-agent system
PyJWT>=2.9.0
apscheduler>=3.10.4
pytest>=8.0.0
pytest-json-report>=1.5.0
